Moved interrupt code to separate file. Added new source file to Makefile.am
